♻️ Refactor consensus
