+++ plugins/consensus/statement.go
@@ -1,10 +1,6 @@
-package statement
+package consensus
 
 import (
-	"context"
-	"time"
-
-	"github.com/iotaledger/goshimmer/dapps/valuetransfers"
 	"github.com/iotaledger/goshimmer/dapps/valuetransfers/packages/transaction"
 	"github.com/iotaledger/goshimmer/packages/tangle"
 	"github.com/iotaledger/goshimmer/packages/vote"
@@ -12,35 +8,6 @@ import (
 	"github.com/iotaledger/goshimmer/plugins/issuer"
 	"github.com/iotaledger/hive.go/identity"
 )
-
-// OpinionGiver is a wrapper for both statements and peers.
-type OpinionGiver struct {
-	view *statement.View
-	pog  *valuetransfers.PeerOpinionGiver
-}
-
-// Query retrievs the opinions about the given conflicts and timestamps.
-func (o *OpinionGiver) Query(ctx context.Context, conflictIDs []string, timestampIDs []string) (opinions vote.Opinions, err error) {
-	for i := 0; i < waitForStatement; i++ {
-		opinions, err = o.view.Query(ctx, conflictIDs, timestampIDs)
-		if err == nil {
-			return opinions, nil
-		}
-		time.Sleep(time.Second)
-	}
-
-	opinions, err = o.pog.Query(ctx, conflictIDs, timestampIDs)
-	if err == nil {
-		return opinions, nil
-	}
-
-	return nil, err
-}
-
-// ID returns a string representation of the identifier of the underlying Peer.
-func (o *OpinionGiver) ID() string {
-	return o.pog.ID()
-}
 
 func makeStatement(roundStats *vote.RoundStats) {
 	// TODO: add check for Mana threshold
@@ -121,7 +88,7 @@ func readStatement(cachedMessageEvent *tangle.CachedMessageEvent) {
 	// TODO: check reduced version VS full
 	issuerID := identity.NewID(solidMessage.IssuerPublicKey()).String()
 
-	issuerRegistry := Registry().NodeRegistry(issuerID)
+	issuerRegistry := Registry().NodeView(issuerID)
 
 	for _, conflict := range statementPayload.Conflicts {
 		issuerRegistry.AddConflict(conflict)
